% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSigTest.R
\name{GeneSigTest}
\alias{GeneSigTest}
\title{Perform Gene Expression Significance Test}
\usage{
GeneSigTest(
  object,
  nMC = 1000,
  p_threshold = 0.05,
  refine = FALSE,
  auto = TRUE,
  filtering = FALSE
)
}
\arguments{
\item{object}{an ENIGMA object}

\item{nMC}{the number of bootstrapping times}

\item{p_threshold}{the threshold of pvalue to determine the significant gene}

\item{refine}{whether perform refinement}

\item{auto}{automatically determine whether perform refinement}

\item{filtering}{perform gene filtering for enigma object. Default: TRUE}
}
\value{
A list object contain two object
 call: A binary matrix (row: gene, column: cell type) indicating which gene has high confidence to be accurated estimated
 pval: A p-value matrix (row: gene, column: cell type) indicating the confidence
}
\description{
Perform Gene Expression Significance Test
}
